services/users/src/requirements.txt
django
pillow
psycopg2-binary
python-dotenv
django-cors-headers
gunicorn
djangorestframework
django-filter
djangorestframework-simplejwt